aplit out actions
