+++ apps/api-journeys/src/modules/action/index.ts
@@ -1,0 +1,1 @@
+export { actionModule } from './action'
